requirements.txt
requests==2.25.0
PyExecJS==1.5.1
beautifulsoup4==4.9.3
